Fix wrong url pattern for Yieldlab (#22)
The domain 'ad.yieldlab.net' is used for all our calls. So if there is a delivery / matching / tracking done on the website, the latency is added up for all the calls and the addon gives the wrong impression that yieldlab's header bidding took a long time to respond. "ad.yieldlab.net/yp/" is the actual endpoint for solely the header bidding.